# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Enterprise API for EvalHub platform with role-based security",
        "title": "EvalHub API",
        "termsOfService": "https://evalhub.com/terms",
        "contact": {
            "name": "EvalHub API Support",
            "url": "https://evalhub.com/support",
            "email": "api-support@evalhub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session-based authentication cookie",
            "type": "apiKey",
            "name": "session_token",
            "in": "cookie"
        }
    },
    "paths": {
...
        "/posts": {
            "get": {
                "security": [
                    {"BearerAuth": []},
                    {"SessionAuth": []}
                ],
                "description": "Retrieves a paginated list of posts",
                "tags": ["Posts"],
                "summary": "List posts",
                "parameters": [
                    {
                        "type": "integer",
                        "minimum": 1,
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100,
                        "default": 20,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "schema": {
                            "$ref": "#/definitions/PostsListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {"BearerAuth": []},
                    {"SessionAuth": []}
                ],
                "description": "Creates a new post",
                "tags": ["Posts"],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Post creation data",
                        "name": "createPostRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "schema": {
                            "$ref": "#/definitions/PostResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "error": {
                    "type": "string",
                    "example": "An error occurred"
                },
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_123456789"
                }
            }
        },
...
        "UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "enum": ["expert", "evaluator", "admin"],
                    "example": "expert"
                }
            }
        },
...
        "CreatePostRequest": {
            "type": "object",
            "required": [
                "title",
                "content",
                "category"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "example": "How to implement authentication in Go"
                },
                "content": {
                    "type": "string",
                    "minLength": 1,
                    "example": "This is the content of the post explaining authentication implementation..."
                },
                "category": {
                    "type": "string",
                    "example": "programming"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["go", "authentication", "jwt"]
                },
                "is_featured": {
                    "type": "boolean",
                    "default": false,
                    "example": false
                }
            }
        },
...
        "PostResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "How to implement authentication in Go"
                },
                "content": {
                    "type": "string",
                    "example": "This is the content of the post..."
                },
                "category": {
                    "type": "string",
                    "example": "programming"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["go", "authentication", "jwt"]
                },
                "author": {
                    "$ref": "#/definitions/UserInfo"
                },
                "is_featured": {
                    "type": "boolean",
                    "example": false
                },
                "likes_count": {
                    "type": "integer",
                    "example": 25
                },
                "dislikes_count": {
                    "type": "integer",
                    "example": 2
                },
                "comments_count": {
                    "type": "integer",
                    "example": 12
                },
                "views_count": {
                    "type": "integer",
                    "example": 156
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T12:00:00Z"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "PostsListResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PostResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/PaginationMeta"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_123456789"
                }
            }
        },
...
        "PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "format": "int64",
                    "example": 20
                },
                "total_pages": {
                    "type": "integer",
                    "format": "int64",
                    "example": 5
                },
                "total_count": {
                    "type": "integer",
                    "format": "int64",
                    "example": 95
                },
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_prev": {
                    "type": "boolean",
                    "example": false
                }
            }
        }
    },
...
}